+++ packages/material-composer-r3f/src/lib/use-detect-shallow-change.tsx
